backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
asyncpg
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
pytest
pytest-asyncio
numpy
pandas
scikit-learn
joblib
deap
openai
langchain
pulp